+++ src/main/java/com/example/namo2/domain/moim/application/MoimScheduleFacade.java
@@ -1,0 +1,20 @@
+package com.example.namo2.domain.moim.application;
+
+import com.example.namo2.domain.moim.application.converter.MoimScheduleConverter;
+import com.example.namo2.domain.moim.application.impl.MoimScheduleService;
+import com.example.namo2.domain.moim.application.impl.MoimService;
+import com.example.namo2.domain.moim.domain.Moim;
+import com.example.namo2.domain.moim.domain.MoimSchedule;
+import com.example.namo2.domain.moim.ui.dto.MoimScheduleRequest;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
+
+@Component
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class MoimScheduleFacade {
+    private final MoimService moimService;
+
+
+}
