+++ src/main/java/com/example/namo2/domain/moim/application/impl/MoimScheduleService.java
@@ -1,0 +1,11 @@
+package com.example.namo2.domain.moim.application.impl;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class MoimScheduleService {
+}
