+++ src/main/java/com/example/namo2/domain/moim/ui/MoimScheduleController.java
@@ -1,0 +1,87 @@
+package com.example.namo2.domain.moim.ui;
+
+import com.example.namo2.domain.moim.application.MoimFacade;
+import com.example.namo2.domain.moim.application.impl.MoimService;
+import com.example.namo2.domain.moim.ui.dto.MoimScheduleDto;
+import com.example.namo2.domain.moim.ui.dto.MoimScheduleRequest;
+import com.example.namo2.global.common.response.BaseResponse;
+import com.example.namo2.global.utils.Converter;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+@Tag(name = "Moim", description = "모임 관련 API")
+@Slf4j
+@RequiredArgsConstructor
+@RestController
+@RequestMapping("/moims/schedule")
+public class MoimScheduleController {
+    private final MoimFacade moimFacade;
+    private final MoimService moimService;
+    private final Converter converter;
+
+    @Operation(summary = "모임 스케쥴 생성", description = "모임 스케쥴 생성 API")
+    @PostMapping("/schedule")
+    public BaseResponse<Long> createMoimSchedule(@Valid @RequestBody MoimScheduleRequest.PostMoimScheduleDto scheduleReq) {
+        Long scheduleId = moimService.createSchedule(scheduleReq);
+        return new BaseResponse(scheduleId);
+    }
+
+    @Operation(summary = "모임 스케쥴 생성", description = "모임 스케쥴 생성 API")
+    @PatchMapping("/schedule")
+    public BaseResponse<Long> modifyMoimSchedule(@Valid @RequestBody MoimScheduleRequest.PatchMoimScheduleDto scheduleReq) {
+        moimService.updateSchedule(scheduleReq);
+        return BaseResponse.ok();
+    }
+
+    @Operation(summary = "모임 스케쥴 카테고리 수정", description = "모임 스케쥴 카테고리 수정 API")
+    @PatchMapping("/schedule/category")
+    public BaseResponse<Long> modifyMoimScheduleCategory(@Valid @RequestBody MoimScheduleRequest.PatchMoimScheduleCategoryDto scheduleReq, HttpServletRequest request) {
+        moimService.updateScheduleCategory(scheduleReq, (Long) request.getAttribute("userId"));
+        return BaseResponse.ok();
+    }
+
+    @Operation(summary = "모임 스케쥴 삭제", description = "모임 스케쥴 삭제 API")
+    @DeleteMapping("/schedule/{moimScheduleId}")
+    public BaseResponse<Long> removeMoimSchedule(@PathVariable Long moimScheduleId) {
+        moimService.deleteSchedule(moimScheduleId);
+        return BaseResponse.ok();
+    }
+
+    @Operation(summary = "월간 모임 스케쥴 조회", description = "월간 모임 스케쥴 조회 API")
+    @GetMapping("/schedule/{moimId}/{month}")
+    public BaseResponse<MoimScheduleDto> getMoimSchedules(@PathVariable("moimId") Long moimId,
+                                                          @PathVariable("month") String month) {
+        List<LocalDateTime> localDateTimes = converter.convertLongToLocalDateTime(month);
+        List<MoimScheduleDto> schedules = moimService.findMoimSchedules(moimId, localDateTimes);
+        return new BaseResponse(schedules);
+    }
+
+    @Operation(summary = "모임 스케쥴 생성 알람", description = "모임 스케쥴 생성 알람 API")
+    @PostMapping("/schedule/alarm")
+    public BaseResponse createMoimScheduleAlarm(@Valid @RequestBody MoimScheduleRequest.PostMoimScheduleAlarmDto postMoimScheduleAlarmDto) {
+        moimService.createScheduleAlarm(postMoimScheduleAlarmDto);
+        return BaseResponse.ok();
+    }
+
+    @Operation(summary = "모임 스케쥴 변경 알람", description = "모임 스케쥴 변경 알람 API")
+    @PatchMapping("/schedule/alarm")
+    public BaseResponse updateMoimScheduleAlarm(@Valid @RequestBody MoimScheduleRequest.PostMoimScheduleAlarmDto postMoimScheduleAlarmDto) {
+        moimService.updateScheduleAlarm(postMoimScheduleAlarmDto);
+        return BaseResponse.ok();
+    }
+}
